+++ changelog.md
@@ -1,2 +1,6 @@
+## 0.2.0
+- Fix Better Combat two-handed weapons not having the "Two-Handed" tooltip line
+- Color enchantment names based on their relation to the maximum enchantment value
+
 ## 0.1.0
 - Initial Release